created a separate module, Data.Parameterized.NatRepr.Internal, which contains the definition of NatRepr and everything required for its instances, and exports all of this. This module is not exposed to the end user, but is needed within Data.Parameterized.Vector.
